// generated/interface/models/InterfaceInfoAddRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type InterfaceInfoAddRequest = {
    description?: string;
    method?: string;
    name?: string;
    requestParam?: string;
    responseBody?: string;
    status?: number;
    url?: string;
};