// generated/interface/models/InterfaceInfoUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type InterfaceInfoUpdateRequest = {
    description?: string;
    id?: number;
    method?: string;
    name?: string;
    requestParam?: string;
    responseBody?: string;
    status?: number;
    url?: string;
};